requirements.txt
passlib
chromadb
sentence-transformers
streamlit
python-jose
bcrypt
motor
dotenv
fastapi
uvicorn
python-multipart